pyproject.toml
[tool.poetry]
name = "em-jan-cafe"
version = "0.1.0"
description = "EM january test task"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.6"
djangorestframework = "^3.15.2"
environs = "^14.1.1"
drf-yasg = "^1.21.8"
pytest-django = "^4.10.0"
coverage = "^7.6.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
